refactor(ruby): split test-DSL and symbols into tests.go/symbols.go
Move the RSpec/test-block DSL to tests.go and class/method emission to
symbols.go. Decompose buildTestScopeSegment into describeScopeSegment/
itScopeSegment helpers and extract emitInheritanceEdges from
handleClassOrModule. Behavior-preserving: all 24 goldens byte-identical.
