+++ internal/extract/ruby/ruby.go
@@ -34,7 +34,6 @@
 package ruby
 
 import (
-	"fmt"
 	"slices"
 	"strings"
 
@@ -238,26 +237,6 @@ func (w *walker) dispatchCall(n *sitter.Node, scope []string) (bool, error) {
 	return false, nil
 }
 
-// rspecDSLMethods is the set of RSpec DSL method names that create
-// test scopes when called with a block.
-var rspecDSLMethods = map[string]bool{
-	"it": true, "describe": true, "context": true,
-	"before": true, "after": true, "around": true,
-	"let": true, "expect": true,
-}
-
-// rspecMatcherMethods is the set of RSpec matcher chain methods that
-// should not be emitted as calls edges — they are DSL sugar, not
-// application method invocations.
-var rspecMatcherMethods = map[string]bool{
-	"to": true, "not_to": true, "to_not": true,
-	"eq": true, "be": true, "be_nil": true, "be_empty": true,
-	"be_valid": true, "be_present": true, "be_a": true,
-	"raise_error": true, "change": true, "receive": true,
-	"have_received": true, "match": true, "include": true,
-	"contain_exactly": true, "start_with": true, "end_with": true,
-}
-
 func (w *walker) walkChildren(n *sitter.Node, scope []string) error {
 	count := n.NamedChildCount()
 	for i := uint(0); i < count; i++ {
@@ -266,195 +245,6 @@ func (w *walker) walkChildren(n *sitter.Node, scope []string) error {
 		}
 	}
 	return nil
-}
-
-// handleTestBlock processes an RSpec DSL call that has a block body.
-// It builds a synthetic scope name, walks the block for calls/identifiers,
-// and recurses into nested test blocks.  Returns true to signal the node
-// was consumed.
-func (w *walker) handleTestBlock(n *sitter.Node, scope []string, methodName string) (bool, error) {
-	// Emit the conventional tests edge for top-level describe with a
-	// constant argument (e.g. `describe Order do … end`).
-	if methodName == "describe" && len(scope) == 0 {
-		if err := w.emitDescribeEdge(n); err != nil {
-			return true, err
-		}
-	}
-
-	block := getBlockChild(n)
-	if block == nil {
-		// No block — but the arguments may contain calls we still want to
-		// capture (e.g. `expect(TopicCreator.create(...))`).
-		synthetic := w.buildSyntheticSource(scope)
-		return true, extract.WalkNamedDescendants(n, "call", func(c *sitter.Node) error {
-			if w.isInsideNestedTestBlock(c, n) {
-				return nil
-			}
-			methodNode := c.ChildByFieldName("method")
-			if methodNode != nil && rspecDSLMethods[extract.Text(methodNode, w.source)] {
-				return nil
-			}
-			return w.emitTestCall(c, synthetic, scope)
-		})
-	}
-
-	segment := w.buildTestScopeSegment(n, methodName)
-	if segment == "" {
-		// Unnamed / unresolvable block — fall back to file-level scope.
-		return true, w.walkTestBlockWithFallback(block, scope)
-	}
-
-	// Push segment and cap depth at 3.
-	w.testScope = append(w.testScope, segment)
-	if len(w.testScope) > 3 {
-		w.testScope = w.testScope[:len(w.testScope)-1]
-		return true, w.walkTestBlockWithFallback(block, scope)
-	}
-	defer func() {
-		w.testScope = w.testScope[:len(w.testScope)-1]
-	}()
-
-	synthetic := w.buildSyntheticSource(scope)
-	body := getBlockBody(block)
-	if body == nil {
-		return true, nil
-	}
-	return true, w.walkTestBody(body, scope, synthetic)
-}
-
-// buildTestScopeSegment extracts a scope segment from a test DSL call.
-// Returns "" when the block should fall back to file-level scope.
-//
-//nolint:gocyclo // 27-10: retired by the ruby extractor split
-func (w *walker) buildTestScopeSegment(n *sitter.Node, methodName string) string {
-	args := n.ChildByFieldName("arguments")
-	if args == nil || args.NamedChildCount() == 0 {
-		return ""
-	}
-	first := args.NamedChild(0)
-	if first == nil {
-		return ""
-	}
-
-	switch methodName {
-	case "describe", "context":
-		switch first.Kind() {
-		case "constant", "scope_resolution":
-			return extract.Text(first, w.source)
-		case "string":
-			if hasInterpolation(first) {
-				return ""
-			}
-			desc := extractStringValue(first, w.source)
-			if desc == "" {
-				return ""
-			}
-			if strings.HasPrefix(desc, "#") || strings.HasPrefix(desc, ".") {
-				return desc
-			}
-			return methodName + "_" + sanitizeDesc(desc)
-		default:
-			return ""
-		}
-	case "it":
-		if first.Kind() == "string" {
-			if hasInterpolation(first) {
-				return ""
-			}
-			desc := extractStringValue(first, w.source)
-			if desc == "" {
-				return ""
-			}
-			return "#it_" + sanitizeDesc(desc)
-		}
-		return ""
-	case "before", "after", "around", "let", "expect":
-		// These DSL nodes rarely carry a descriptive string arg;
-		// fall back to file-level scope.
-		return ""
-	default:
-		return ""
-	}
-}
-
-// buildSyntheticSource joins the class/module scope with the test-scope
-// stack into a single synthetic qualified name.
-func (w *walker) buildSyntheticSource(scope []string) string {
-	classScope := strings.Join(scope, "::")
-	testScope := strings.Join(w.testScope, "#")
-
-	if classScope == "" && testScope == "" {
-		return ""
-	}
-	if classScope == "" {
-		return testScope
-	}
-	if testScope == "" {
-		return classScope
-	}
-	return classScope + "#" + testScope
-}
-
-// walkTestBody walks a test block body emitting calls edges with the
-// given synthetic source.  It also recurses into nested test blocks.
-// A single tree walk handles both emission and recursion.
-func (w *walker) walkTestBody(body *sitter.Node, scope []string, source string) error {
-	if err := extract.WalkNamedDescendants(body, "call", func(c *sitter.Node) error {
-		if w.isInsideNestedTestBlock(c, body) {
-			return nil
-		}
-		methodNode := c.ChildByFieldName("method")
-		if methodNode == nil {
-			return nil
-		}
-		methodName := extract.Text(methodNode, w.source)
-		if rspecDSLMethods[methodName] {
-			// Recurse into nested test block.
-			_, err := w.handleTestBlock(c, scope, methodName)
-			return err
-		}
-		return w.emitTestCall(c, source, scope)
-	}); err != nil {
-		return err
-	}
-	// Emit edges for bare identifiers in statement and value positions.
-	return w.emitBareIdentifierCalls(body, source, extract.ConfidenceTests, nil)
-}
-
-// isInsideNestedTestBlock returns true if n sits inside a test DSL call
-// that is a descendant of body (i.e., a nested test block).
-func (w *walker) isInsideNestedTestBlock(n, body *sitter.Node) bool {
-	bodyID := body.Id()
-	for p := n.Parent(); p != nil && p.Id() != bodyID; p = p.Parent() {
-		if p.Kind() == "call" {
-			methodNode := p.ChildByFieldName("method")
-			if methodNode != nil && rspecDSLMethods[extract.Text(methodNode, w.source)] {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// walkTestBlockWithFallback walks a test block body using the file path
-// as the source scope (file-level fallback).
-func (w *walker) walkTestBlockWithFallback(block *sitter.Node, scope []string) error {
-	body := getBlockBody(block)
-	if body == nil {
-		return nil
-	}
-	source := w.fileLevelScope(block)
-	return w.walkTestBody(body, scope, source)
-}
-
-// fileLevelScope returns a file-level synthetic scope like "test.rb#L42".
-func (w *walker) fileLevelScope(n *sitter.Node) string {
-	base := w.filePath
-	if i := strings.LastIndex(base, "/"); i >= 0 {
-		base = base[i+1:]
-	}
-	line := extract.Line(n.StartPosition())
-	return fmt.Sprintf("%s#L%d", base, line)
 }
 
 // getBlockChild returns the block child (do_block or block) of a call node.
@@ -567,201 +357,6 @@ func mergeMaps(base, overlay map[string]string) map[string]string {
 	return result
 }
 
-// hasInterpolation returns true if a string node contains interpolation.
-func hasInterpolation(n *sitter.Node) bool {
-	if n.Kind() != "string" {
-		return false
-	}
-	for i := uint(0); i < n.NamedChildCount(); i++ {
-		c := n.NamedChild(i)
-		if c != nil && c.Kind() == "interpolation" {
-			return true
-		}
-	}
-	return false
-}
-
-// sanitizeDesc turns a human-readable block description into a safe
-// scope segment: spaces → underscores, strip non-alphanumerics.
-func sanitizeDesc(s string) string {
-	var b strings.Builder
-	for _, r := range s {
-		switch {
-		case r >= 'a' && r <= 'z', r >= 'A' && r <= 'Z', r >= '0' && r <= '9':
-			b.WriteRune(r)
-		case r == ' ':
-			b.WriteRune('_')
-		default:
-			// Drop punctuation.
-		}
-	}
-	return b.String()
-}
-
-// handleClassOrModule emits the symbol, records inheritance (class only),
-// and descends into the body with the class/module pushed onto the scope.
-//
-//nolint:gocognit // 27-10: retired by the ruby extractor split
-func (w *walker) handleClassOrModule(n *sitter.Node, scope []string, kind model.SymbolKind) error {
-	nameNode := n.ChildByFieldName("name")
-	if nameNode == nil {
-		return w.walkChildren(n, scope) // malformed, keep walking
-	}
-	name := extract.Text(nameNode, w.source)
-	if name == "" {
-		return w.walkChildren(n, scope)
-	}
-
-	// A::B as a class name pushes both segments; the last segment is the
-	// "name" (what grep will find), the full chain is the qualified name.
-	segments := strings.Split(name, "::")
-	newScope := append(slices.Clone(scope), segments...)
-	qualified := strings.Join(newScope, "::")
-	parent := strings.Join(scope, "::")
-
-	if err := w.emit.Symbol(extract.EmittedSymbol{
-		Name:            segments[len(segments)-1],
-		Qualified:       qualified,
-		Kind:            kind,
-		ParentQualified: parent,
-		LineStart:       extract.Line(n.StartPosition()),
-		LineEnd:         extract.Line(n.EndPosition()),
-		Docstring:       docstringFor(n, w.source),
-	}); err != nil {
-		return err
-	}
-
-	// Inheritance: emit an edge when the superclass is a simple constant.
-	// Target resolution to a symbol_id happens at write time — here we
-	// just record the target's qualified name.
-	if kind == model.KindClass {
-		if sup := n.ChildByFieldName("superclass"); sup != nil {
-			if target := superclassName(sup, w.source); target != "" {
-				line := extract.Line(sup.StartPosition())
-				if err := w.emit.Edge(extract.EmittedEdge{
-					SourceQualified: qualified,
-					TargetQualified: target,
-					Kind:            model.EdgeInherits,
-					Line:            &line,
-					Confidence:      1.0,
-				}); err != nil {
-					return err
-				}
-				if isTestSuperclass(target) {
-					if testedClass := inferTestedClass(qualified); testedClass != "" {
-						if err := w.emit.Edge(extract.EmittedEdge{
-							SourceQualified: qualified,
-							TargetQualified: testedClass,
-							Kind:            model.EdgeTests,
-							Line:            &line,
-							Confidence:      extract.ConfidenceConvention,
-						}); err != nil {
-							return err
-						}
-					}
-				}
-			}
-		}
-	}
-
-	if body := n.ChildByFieldName("body"); body != nil {
-		ivarTypes := buildInstanceVarTypeMap(body, w.source)
-		w.classInstanceVars[qualified] = ivarTypes
-		// Pre-pass: record each direct instance method's visibility before the
-		// methods are emitted, so handleMethod can attach it. Mirrors the
-		// ivar-type pre-pass above.
-		w.recordBodyVisibility(body, qualified)
-		if err := w.walkChildren(body, newScope); err != nil {
-			return err
-		}
-		// Capture references that live at class-body level rather than
-		// inside a method: constant RHS values, and exception/handler
-		// classes named by DSLs like `rescue_from Foo` / `retry_on Bar`.
-		// Method-body references are attributed to their method, so they
-		// are skipped here via the nested-def guard.
-		return w.collectConstRefs(body, qualified, true)
-	}
-	return nil
-}
-
-// handleMethod emits a method symbol qualified either as Class#name
-// (instance) or Class.name (singleton). For top-level methods the
-// separator and parent are both empty — they become KindMethod with
-// qualified=name, which matches how Ruby treats top-level defs (they
-// get attached to Object at runtime, but we don't model Object here).
-//
-// After emitting, the body is walked for call nodes so intra-body
-// calls land as calls edges.
-func (w *walker) handleMethod(n *sitter.Node, scope []string, singleton bool) error {
-	nameNode := n.ChildByFieldName("name")
-	if nameNode == nil {
-		return nil
-	}
-	name := extract.Text(nameNode, w.source)
-	if name == "" {
-		return nil
-	}
-
-	parent := strings.Join(scope, "::")
-	var qualified, receiver string
-	switch {
-	case parent == "":
-		// Top-level def: attaches to Object at runtime; no class-vs-instance
-		// dispatch distinction to record.
-		qualified = name
-	case singleton:
-		qualified = parent + "." + name
-		receiver = extract.ReceiverSingleton
-	default:
-		qualified = parent + "#" + name
-		receiver = extract.ReceiverInstance
-	}
-
-	// Visibility comes from the per-class pre-pass (instance methods only).
-	// Anything not recorded — singleton methods, top-level defs, methods the
-	// pre-pass could not classify — defaults to public, which is the safe
-	// direction: a public method can never earn a `dead` verdict.
-	visibility := "public"
-	if v, ok := w.methodVisibility[qualified]; ok {
-		visibility = v
-	}
-
-	if err := w.emit.Symbol(extract.EmittedSymbol{
-		Name:            name,
-		Qualified:       qualified,
-		Kind:            model.KindMethod,
-		Receiver:        receiver,
-		Visibility:      visibility,
-		ParentQualified: parent,
-		LineStart:       extract.Line(n.StartPosition()),
-		LineEnd:         extract.Line(n.EndPosition()),
-		Docstring:       docstringFor(n, w.source),
-	}); err != nil {
-		return err
-	}
-	// Include / extend / prepend calls inside a method body come through
-	// here too and are emitted as regular calls — the `includes` edges
-	// are produced separately at the class-body level in handleIncludeCall,
-	// and a dynamic include at runtime is rare enough that a bare calls
-	// edge is an accurate record of what was written.
-	body := n.ChildByFieldName("body")
-	localTypes := buildLocalTypeMap(body, w.source)
-	w.addRescueBindings(body, localTypes)
-	ivarTypes := w.classInstanceVars[strings.Join(scope, "::")]
-	if err := extract.WalkNamedDescendants(body, "call", func(c *sitter.Node) error {
-		if isInsideBlock(c) {
-			return nil
-		}
-		return w.emitCall(c, qualified, scope, localTypes, ivarTypes)
-	}); err != nil {
-		return err
-	}
-	if err := w.emitBareIdentifierCalls(body, qualified, extract.ConfidenceDynamic, methodParamNames(n, w.source)); err != nil {
-		return err
-	}
-	return w.collectConstRefs(body, qualified, false)
-}
-
 // emitCall produces a calls edge for one `call` node. The target is
 // resolved from the receiver when possible: `self` and implicit calls
 // are emitted as `self.name` so the resolver rewrites them to the
@@ -818,19 +413,6 @@ func (w *walker) emitCall(n *sitter.Node, source string, scope []string, localTy
 		return nil
 	}
 	return w.emitCallWithConfidence(n, source, scope, localTypes, ivarTypes, confidence)
-}
-
-// emitTestCall delegates to emitCall but substitutes ConfidenceTests and
-// skips RSpec matcher noise (eq, be_valid, raise_error, etc.).
-func (w *walker) emitTestCall(n *sitter.Node, source string, scope []string) error {
-	methodNode := n.ChildByFieldName("method")
-	if methodNode == nil {
-		return nil
-	}
-	if rspecMatcherMethods[extract.Text(methodNode, w.source)] {
-		return nil
-	}
-	return w.emitCallWithConfidence(n, source, scope, nil, nil, extract.ConfidenceTests)
 }
 
 // emitCallWithConfidence is emitCall's core logic with an injectable
@@ -2313,49 +1895,6 @@ func (w *walker) emitScopeEdge(n *sitter.Node, scope []string) error {
 	})
 }
 
-// emitDescribeEdge detects RSpec.describe/describe with a constant
-// argument and emits a tests edge to the referenced class.
-func (w *walker) emitDescribeEdge(n *sitter.Node) error {
-	// For RSpec.describe, the receiver is "RSpec". For bare describe, no receiver.
-	// Both are valid — just need the first arg to be a constant.
-	if recv := n.ChildByFieldName("receiver"); recv != nil {
-		if extract.Text(recv, w.source) != "RSpec" {
-			return nil
-		}
-	}
-
-	args := n.ChildByFieldName("arguments")
-	if args == nil || args.NamedChildCount() == 0 {
-		return nil
-	}
-	first := args.NamedChild(0)
-	if first == nil {
-		return nil
-	}
-
-	var target string
-	switch first.Kind() {
-	case "constant":
-		target = extract.Text(first, w.source)
-	case "scope_resolution":
-		target = extract.Text(first, w.source)
-	default:
-		return nil
-	}
-	if target == "" {
-		return nil
-	}
-
-	line := extract.Line(n.StartPosition())
-	return w.emit.Edge(extract.EmittedEdge{
-		SourceQualified: target + "Test",
-		TargetQualified: target,
-		Kind:            model.EdgeTests,
-		Line:            &line,
-		Confidence:      extract.ConfidenceConvention,
-	})
-}
-
 // resourceActions is the set of RESTful actions for `resources`.
 var resourceActions = []string{"index", "show", "new", "create", "edit", "update", "destroy"}
 
@@ -2665,37 +2204,6 @@ func extractStringValue(n *sitter.Node, source []byte) string {
 			if c != nil && c.Kind() == "string_content" {
 				return extract.Text(c, source)
 			}
-		}
-	}
-	return ""
-}
-
-// isTestSuperclass returns true if the superclass name indicates a test base class.
-func isTestSuperclass(name string) bool {
-	return strings.Contains(name, "TestCase") || strings.Contains(name, "IntegrationTest") || strings.Contains(name, "SystemTest")
-}
-
-// inferTestedClass strips "Test" suffix from a class name to find what it tests.
-// "UserTest" → "User", "Admin::DashboardControllerTest" → "Admin::DashboardController"
-func inferTestedClass(qualified string) string {
-	if strings.HasSuffix(qualified, "Test") {
-		return strings.TrimSuffix(qualified, "Test")
-	}
-	return ""
-}
-
-// superclassName pulls the target class name from a `superclass` node.
-// The node wraps its target (usually `constant` or `scope_resolution`).
-func superclassName(sup *sitter.Node, source []byte) string {
-	count := sup.NamedChildCount()
-	for i := uint(0); i < count; i++ {
-		c := sup.NamedChild(i)
-		if c == nil {
-			continue
-		}
-		switch c.Kind() {
-		case "constant", "scope_resolution":
-			return c.Utf8Text(source)
 		}
 	}
 	return ""
